TJFE-52 [rename] common 패키지 경로 변경( presentation/ui/common -> presentation/common)
