+++ presentation/src/main/java/com/tedmoon99/presentation/MainActivity.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.tedmoon99.presentation.ui.common.navigation.MainNavigation
-import com.tedmoon99.presentation.ui.common.theme.TravelJournalForEveryoneTheme
+import com.tedmoon99.presentation.common.navigation.MainNavigation
+import com.tedmoon99.presentation.common.theme.TravelJournalForEveryoneTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
