+++ presentation/src/main/java/com/tedmoon99/presentation/common/navigation/MainNavigation.kt
@@ -1,11 +1,11 @@
-package com.tedmoon99.presentation.ui.common.navigation
+package com.tedmoon99.presentation.common.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.tedmoon99.presentation.home.HomeScreen
-import com.tedmoon99.presentation.ui.common.screen.Screen
+import com.tedmoon99.presentation.common.screen.Screen
 
 @Composable
 fun MainNavigation() {
