+++ presentation/src/main/java/com/tedmoon99/presentation/common/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.tedmoon99.presentation.ui.common.theme
+package com.tedmoon99.presentation.common.theme
 
 import androidx.compose.ui.graphics.Color
