+++ presentation/src/main/java/com/tedmoon99/presentation/common/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.tedmoon99.presentation.ui.common.theme
+package com.tedmoon99.presentation.common.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
